Improvements
1. Bottom navigation bar
2. Quiz section
