+++ android/app/build.gradle
@@ -70,4 +70,5 @@ dependencies {
     implementation 'com.google.firebase:firebase-crashlytics:17.2.2'
     implementation(platform("com.google.firebase:firebase-bom:32.2.3"))
     implementation("androidx.multidex:multidex:2.0.1")
+    implementation "dev.shreyaspatil.EasyUpiPayment:EasyUpiPayment:3.0.3"
 }
