+++ lib/home/home_screen.dart
@@ -1,11 +1,13 @@
-import 'package:animated_bottom_navigation_bar/animated_bottom_navigation_bar.dart';
 import 'package:carousel_slider/carousel_slider.dart';
+import 'package:fancy_bottom_navigation_2/fancy_bottom_navigation.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import '../app_utils/color_constants.dart';
 import '../app_utils/custom_carousel.dart';
 import '../app_utils/image_utils.dart';
 import 'home_controller.dart';
+import 'ui_components/quiz_section/index.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({Key? key}) : super(key: key);
@@ -76,8 +78,8 @@ class HomeScreen extends StatelessWidget {
                                 ),
                                 const SizedBox(height: 24),
                                 CustomCarousel(
-                                  boxHgt: 190,
-                                  boxWdt: Get.width * 0.95,
+                                  boxHgt: 170,
+                                  boxWdt: double.maxFinite,
                                   itemList: controller.runningTournament,
                                   index: controller.runningIndex,
                                   carouselController:
@@ -107,73 +109,7 @@ class HomeScreen extends StatelessWidget {
                           )
                         ],
                       ),
-                      Container(
-                        width: double.maxFinite,
-                        height: 150,
-                        margin: const EdgeInsets.only(top: 16),
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 20,
-                          vertical: 10,
-                        ),
-                        decoration: const BoxDecoration(
-                          color: Colors.white,
-                          borderRadius: BorderRadius.only(
-                            bottomLeft: Radius.circular(32),
-                            bottomRight: Radius.circular(32),
-                          ),
-                          boxShadow: [
-                            BoxShadow(
-                              color: Colors.black26,
-                              offset: Offset(0, 1),
-                              spreadRadius: 2,
-                              blurRadius: 2,
-                            ),
-                          ],
-                        ),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          mainAxisAlignment: MainAxisAlignment.spaceAround,
-                          children: [
-                            Center(
-                              child: Text(
-                                controller.questionList[0].Question,
-                                style: TextStyle(
-                                  fontSize: 16,
-                                  fontWeight: FontWeight.w500,
-                                  color: ColorConstants().redColor,
-                                ),
-                              ),
-                            ),
-                            const SizedBox(height: 8),
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              children: [
-                                AnswerWgt(
-                                  txt: controller.questionList[0].Option1,
-                                  isSelected: controller.selectedAns == 0,
-                                ),
-                                AnswerWgt(
-                                  txt: controller.questionList[0].Option2,
-                                  isSelected: controller.selectedAns == 1,
-                                ),
-                              ],
-                            ),
-                            Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              children: [
-                                AnswerWgt(
-                                  txt: controller.questionList[0].Option3,
-                                  isSelected: controller.selectedAns == 3,
-                                ),
-                                AnswerWgt(
-                                  txt: controller.questionList[0].Option4,
-                                  isSelected: controller.selectedAns == 4,
-                                ),
-                              ],
-                            ),
-                          ],
-                        ),
-                      ),
+                      QuizSection(controller: controller),
                       const Padding(
                         padding: EdgeInsets.only(
                           top: 24,
@@ -191,7 +127,7 @@ class HomeScreen extends StatelessWidget {
                         ),
                       ),
                       CustomCarousel(
-                        boxHgt: 190,
+                        boxHgt: 170,
                         boxWdt: Get.width * 0.95,
                         itemList: controller.upComingTournament,
                         index: controller.upComingIndex,
@@ -219,7 +155,7 @@ class HomeScreen extends StatelessWidget {
                         ),
                       ),
                       CustomCarousel(
-                        boxHgt: 190,
+                        boxHgt: 170,
                         boxWdt: Get.width * 0.95,
                         itemList: controller.runningTournament,
                         index: controller.runningIndex,
@@ -250,7 +186,7 @@ class HomeScreen extends StatelessWidget {
                       ),
 
                       SizedBox(
-                        height: 190,
+                        height: 170,
                         width: Get.width * 0.95,
                         child: CarouselSlider(
                           items: controller.completeTournament,
@@ -287,7 +223,7 @@ class HomeScreen extends StatelessWidget {
                       ),
 
                       SizedBox(
-                        height: 190,
+                        height: 170,
                         width: Get.width * 0.95,
                         child: CarouselSlider(
                           items: controller.cancelTournament,
@@ -307,73 +243,16 @@ class HomeScreen extends StatelessWidget {
                   ),
                 ),
               ),
-        floatingActionButton: const Icon(Icons.shopping_cart),
-        floatingActionButtonLocation:
-            FloatingActionButtonLocation.miniCenterDocked,
-        bottomNavigationBar: AnimatedBottomNavigationBar(
-          icons: controller.iconList,
-          activeIndex: 0,
-          activeColor: Colors.green,
-          gapLocation: GapLocation.center,
-          notchSmoothness: NotchSmoothness.verySmoothEdge,
-          leftCornerRadius: 32,
-          rightCornerRadius: 32,
-          onTap: (index) {},
+        bottomNavigationBar: FancyBottomNavigation(
+          tabs: [
+            TabData(iconData: Icons.home_outlined, title: "Home"),
+            TabData(iconData: Icons.live_tv_outlined, title: "Live"),
+            TabData(iconData: Icons.shopping_cart_outlined, title: "Cart"),
+            TabData(iconData: CupertinoIcons.photo, title: "Gallety"),
+            TabData(iconData: Icons.person_3_rounded, title: "Profile"),
+          ],
+          onTabChangedListener: (position) => controller.setBottomNavigationBarSelectedItem(position)
         ),
-      ),
-    );
-  }
-}
-
-class AnswerWgt extends StatelessWidget {
-  final String txt;
-  final bool isSelected;
-
-  const AnswerWgt({super.key, required this.txt, required this.isSelected});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      height: 35,
-      width: Get.width * 0.4,
-      decoration: BoxDecoration(
-        border: Border.all(
-          color: isSelected
-              ? ColorConstants().greenColor
-              : const Color.fromRGBO(0, 0, 0, 0.5),
-        ),
-        borderRadius: BorderRadius.circular(6),
-        color: isSelected ? ColorConstants().greenColor : Colors.white,
-      ),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: [
-          Text(
-            txt,
-            style: TextStyle(
-              fontSize: 12,
-              fontWeight: FontWeight.w400,
-              color: isSelected
-                  ? ColorConstants().whiteColor
-                  : const Color.fromRGBO(0, 0, 0, 0.5),
-            ),
-          ),
-          Container(
-            width: 18,
-            height: 18,
-            alignment: Alignment.center,
-            decoration: BoxDecoration(
-              shape: BoxShape.circle,
-              color: isSelected ? ColorConstants().whiteColor : Colors.white70,
-              border: Border.all(color: const Color.fromRGBO(0, 0, 0, 0.5)),
-            ),
-            child: Icon(
-              Icons.done,
-              color: isSelected ? ColorConstants().greenColor : Colors.white,
-              size: 12,
-            ),
-          )
-        ],
       ),
     );
   }
