+++ pubspec.yaml
@@ -36,8 +36,10 @@ dependencies:
   carousel_slider: ^4.2.1
   sms_autofill: ^2.3.1
 
-  animated_bottom_navigation_bar: ^1.3.3
+  fancy_bottom_navigation_2: ^0.3.5
   youtube_player_flutter: ^9.0.0
+
+  easy_upi_payment: ^0.1.6
 
 dev_dependencies:
   flutter_test:
